A texture in OpenGL is required to have a width and a height that is a power of 2. So all  textures was pasted on an ARGBFrame frame that was a powerOf2 and the extra texels of the left side and the bottom side were cleared with a value of 0.
There was a bug in the way a new texture was cleared. A border of 10 texels was added after the texture and set to 0. However sometimes that 10 texel offset could exceed the ARGBFrame allocated for the texture and thus cause a heap corruption. This has now been rectified with the piece of code below. That happened in the case of saving to 160 x 120. For the height part, the nearest power of 2 is 128. Only 8 texels higher than the normal height. But we're clearing 10 texels. Hence the crash. Bug number #16308

Code added:


	if( nBorder + width  >= width_pof2 || 
	   nBorder + height >= height_pof2 )
	{
		int nBorder1 =  width_pof2 - width;
		int nBorder2 = height_pof2 - height;
		nBorder = min( nBorder1, nBorder2 ); 
	}