0016348: ShowTime API needs facility to independently control playback volume of sync sound, sound effects and music track.
Added a SetParam API to Show interface to pass on any
persistent rendering information to a Show or its
contained Shows. This mechanism is now used to implement
the audio volume control facility.

Note that this facility is not yet hooked to the UI. It
will get hooked once the constructor generates the audio
visual Shows directly without going via the XML timeline.