Added Dialog class
Dialog class is inherited from the Window control. The purpose of this class/control is to have a non-resizable window with title bar that can be dragged around and close button with customizable style and template. This is useful for creating sub-windows.

The reason why the normal window is not useful is because the title bar disappears when style is set to None.