Correction:
Due the fact that SetPixelFormat() is not working properly in the c# codes (AppWPF\AppWPF.sln), we put it in the constructor of ShowTime::FullScreenRenderer. If the preview window is created by calling mvglCreateWindow() as it is before this change, SetPixelFormat() will be called twice when the FullScreenRenderer object is created, which is not recommended by Windows API. By calling mvCreateWindow() to create the preview window, the problem that MESCore.exe crashes at the end is resolved, which is a result of svn359. However, the reason is not clear. Possibly it is because of the conflict between the two opengl contexts and that SetPixelFormat() is called twice.