0016557: Title/credit preview occupies entire Personalize dialog, instead of preview area
Change ImageTarget's member FSize to FRect to allow drawing in a smaller area within a preview window. Now for title/credit preview, you can see the right preview in the right position, except that the rest of the windowturns black.

Todo:
Sze Lee is trying to create a child window of the title/credit preview main window, where the real preview is shown. If that child window can be created, the issue should be resolved.