Fix: To make muvee preview stop properly when there are video inputs.
"App.mvEngine.SeekPreviewToPos(0.0);" is called in MVMain.xaml.cs when muvee preview stopped. Since seek() of a video is not implemented yet, calling it any time causes problem.

Temp solution: In FullScreenRenderer::Seek(), force it to return after Stop() if the input "pos" < 0.0001.

Todo: Remove "if(pos < 0.0001) return;" when Seek() is implemented.