This revision has a combination of changes.
All objects are set to NULL in SetupOpenGLEffects() and FinalConstruct()

Empty text is not checked by looking at the size() of the string but by looking if content is "" or "{}"


In FinalRelease() we set the objects to NULL after deleting them

We no longer do the .free() for the argbFrame object. That's because argbFrame never created anything so it does not have any responsability for any sub-objects.