FinalConstruct() was misplaced. This error happened while implementing The zoom-in and zoom-out effect.
It's positioning within SetupOpenGLEffect() is critical because it sets the font parameters.

If the case of an animated text, the character height was regularly updated so the problem did not show up.

However in the case of captions where are usually static text, we need to call FinalConstruct() before we set the caracter height.

This has now been rectified.