This is the first batch of commits concerning the replacement of XML generated timelines with Show timelines.
The code is not activated rite now. Inside TextBase.cpp in RenderShow() function, there are two distinct paths for generating shows. the first one is reading the xml auto-generated file and the second on is to read the member variables of textbase and MuseBasedTextStyle directly to create the shows. Only the first path is activated rite now. 

Uncommenting the following line:

//ShowTime::AVShow av = CreateShow(); 

will enable the new show creation mechanism.

So far, the only thing has is supported is a picture background with text on top of it. Effects, moos, color layers are not supported.

OpenGLMixers.h now exports the 

	MESAPI static OpenGLTx *Create(FString *effectName);

function.

MuseBasedStyle.h/.cpp has two functions CreateTopEffectsShow() and AddEffectsToShow(). they dont really do much rite now but that'll soon change.