0016577: Only a small portion of muvee shows in full screen preview
Finally, Kumar detected the problem! It can be understood in this way: when a preview renderer object is created, a player thread <ta> and a rendering context <rc> are created and associated to each other. When a second muvee is made, the old paire is released as the old preview renderer object is destroyed, and then a new preview renderer object is created. With the current driver of nVidia Quadro FX 550/540 cards on Vista, something goes wrong with the association of the new pair of player thread and rendering context and leads to this issue. It could be considered as a driver problem <http://forums.nvidia.com/index.php?showtopic=27043&pid=293761&mode=threaded&show=&st=&#entry293761>. However, we've got a work around solution in this commitment. 

Solution:
We created a new structure <RenderWindowInfo> in FullScreenRenderer and associate it to the OpenGL window <Get/SetWindowLongPtr()> to keep track of the pair of player thread and rendering context. In this case, the pair is created once and live forever till the application quits. So far it's been tested and works well.

Todo:
We need to catch the application close message and release the RenderWindowInfo structure (player thread + rendering context).