Inside VariableParam.h, two functions are now exported from MESCore.dll. They are AnimatedFloatParam(...) and AnimatedStringParam(...).
SetVarStringParam(...) is now exported from within OpenGLMixers.h

CreateShowVideoBackground() has been added to TextBase and as its name suggests, we can not create video backgrounds directly using showtimw without having to depend on the xml parser.

As for string Animations, zoom-in, zoom-out, fade-in, fade-out and the typewriter effects are now implemented inside ApplyTextEffect(..).

That's it for this check-in. 

The next step is to add background and foreground effects such as rapidOverlay and sepiaTone to titles and credits.