Since the Show_ColorClip class was added to MES, we can now add colorbackgroud support in TextBase. The new function is called
 TextBase::CreateShowColorBackground()

Again since Show_ColorClip class was added to MES, we can now add support for moo animations. moo animations are basically efffects that have been applied to a black color background. The new function is called 

 TextBase::CreateShowMooBackground(..)


Finally audio support was added to titles and credits. The new function is called:

TextBase::AddShowSoundEffects(..)

There is one issue with sound effects though. It can only play mp3 files. Not the mvx files that we ship with styles. This format will have to be supported in the future.


minor change: textdata is now a member variable. This is because it's widely used by all the show functions.