0016992: On machines that don't support Frame Buffer Objects (FBOs), we should fall back on p-buffers.
These changes allow Anchorvale to fall back on basic texture mapping when FBOs are not supported.

New logics:
In the first muvee preview, we determine if FBOs are supported in the target machine. Then,
a. When preview window is small, (small means width<=720 && height<=480), render directly to the preview window without using FBOs;
b. When preview is switched to full screen, check if FOBs are supported:
   YES --> Render to full screen using FBOs, same as old implementation;
   NO  --> Render muvee to a 512X512 texture and then expand the texture to full screen;

Now Anchorvale should be able to run on machines that does not support FBOs due to:
1. Hardware too old;
2. Driver too old;