0016992: On machines that don't support Frame Buffer Objects (FBOs), we should fall back on p-buffers. FBOs are no longer used for both muvee preview and saving.
New logics:
For previewing,
a. When preview window is small, (small means width<=720 && height<=480), render directly to the preview window;
b. When preview is switched to full screen, render muvee to a 720X480 texture and then expand the texture to full screen;
For saving,
Use p-buffers if supported. Otherwise use a standard buffer of a hidden window (80X60).