0017164: [TC19-01-01] Application crashes when launching magicMoments 0017187: Crash when trying to view any picture 0017156: Rotate picture crashes
Changing the acceleration/deceleration ratio from 0.1/0.9 to 0.3/0.7 so that the app doesn't crash using armed mvAppSKDL.dll on Vista

According to MSDN, the sum of acceleration and deceleration must be equal or less than 1.0

Gerry has pointed out that perhaps the calculation of 0.1+0.9 exceeded 1.0 (by a bit..) and causes the crash. It was verified when tested!

So perhaps the armed mvAppSDK causes the double calculation to deviate slightly.