0017255: Number of pictures loaded ahead needs to be adjusted according to GPU memory
The use of <DirectDraw> is taken out;
Use <DxDiag> to get the size of dedicated video memory;
Wrap everything in <VRam.h>;
<int GetVRam()> is the global function in <namespace ShowTime> that does the job;

Now a static call to query the size of dedicated video memory will be made in the constructor of <RenderEnvBase>. A delay of 1~2 seconds may occur due to this call, which means it's slower in the very first time you preview/save a muvee after Anchorvale is launched every time.

The logic on low VRam machines keeps the same as it's mentioned in Rev1344.