0017584: New effect framework and pipeline for Anchorvale FEATURE+BUGFIX: Aspect ratio handling was incorrect in earlier versions with respect to style authoring. For example, the style author couldn't rotate a picture by 45 degrees and maintain the correct aspect ratio in the process. These modifications address that issue. So now you can freely combine geometrical transformations such as translation, rotation and scaling with KenBurns on pictures. It works by changing the global coordinate system to [(-AR,-1)-(AR,1)] where AR is the viewport's aspect ratio. When images are drawn using RenderEnv::Force() function, the correct aspect ratio is used so that square things turn out to be square.
TextOut needed to be modified as a consequence of the aspect ratio 
related changes. The modification gets rid of all GLTexture2D 
dependencies in TextOut effect.

BUGFIX: More than 9 images could not be used with the new 
effect framework. This was due to EffectBase not propagating
Leave() calls to its inputs.