0017584: New effect framework and pipeline for Anchorvale
BUGFIX: ColorQuad wasn't compatible with alpha blending.
It was correctly taking alpha into account, but would have ended up
in the wrong z-order if it was used in conjunction with transforms.
The right way to handle the case where the color quad needs to
be transparent is to delay the drawing until after all the opaque
elements are drawn and draw it without modifying the depth buffer.
This mechanism is already provided in RenderEnv. ColorQuad
just needs to use it postpone its drawing.