Added the addons' dependency to WPFToolbox
The previous way of adding the class by link has the advantage of not requiring the WPFToolbox.dll during runtime, as it is added statically into the class. However, there is a problem of unable to apply skin from the application level to the addon's UI control, causing inconsistent UI look.

The direct dependency to WPFToolbox may be straightforward, but it requires multiple copies of WPFToolbox.dll (and not to mention their localized resources). Forward and backward compatibility will be an issue too.

The next step would be to move the WPFToolbox to a common area such as GAC (Global Assembly Cache) directory, so that it can be shared.