MANTIS:   	 0017768: Title and Credits muse parameter overrides are being ignored.
museBasedTextStyle used to read the scm file of the styles in cases where there are custom animations that the style author wanted and fed them to an MPEffectParam.

However with the new effects framework, there is a new "Parameter" storage structure.

That means that all title and credits custom parameters were ignored.

Fix:

musedBased Parameter overrides need to be fed to a "Parameter" object.