Changed Release output of WPFToolbox.dll to output\binaries. Also copy to output\binaries\common. (Same goes for the locale dlls.)
Although the output path was bin\Release originally, there is still a copy of WPFToolbox.dll in output\binaries and output\binaries\common due to the side-by-side dependency of AppWPF and SharePlugins. Because WPFtoolbox.dll is installed to GAC now, the WPFToolbox.dll is now no longer copied side-by-side by default.