0017871: [TC16-05-01] 16:9 Video Source in a 16:9 muvee would Result in a 4:3 Squashed Output
The top level RenderFrame() function was rendering a texture given to it
as though it was a square. Modified it to render using the correct aspect ratio.
My local modification to ultra plain as a perspective transform on it which
applies the correct aspect ratio, so I never got to see this problem.