0017601: [TC12-01-04] Zoomed style preview is partially covered in the styles box 0017583: Expanded style preview too big when screen resolution is 1024x768 0017582: Style list doesn't auto-scroll to show expanded style preview 0017225: Tweaks to behavior of big style preview
Few tweaks to Mary's fix.
Set Maximum and Minimum width and height, in addition to the ratio of listboxheight.
This is to ensure that the size of preview is never expanded too big, resulting in poor quality of preview. 
Currently, the min size preview is 100x75, and max preview size is set to 240x180.
Hence if the listbox size allow, double click will expand it to the max 240x180, but not any larger. When listbox size is smaller, it will expand to fit 0.8 of the listbox's height. The width adjust accordingly to the aspect ratio.

The second parameter of BringIntoView is set to -10 to ensure whole listbox item is visible and not cropped.