0017025: [TC30-04-05] The muvee in the muvee list(muveeShare) have no thumbnail.
BUGFIX: Rgd note 45494 - the STextOut effect did not release its texture in its
Leave() override. Therefore added a Leave override in STextOut. Also put in a
reference counter m_constructionCycle to keep track of the number of times
FinalConstruct() and FinalRelease are called. This got rid of a crash in 
Layer.h when deleting a bitmap.