0018149: S00509_MusicVideo3: Stack overflow in muse_cell _celli
DEBUGGING: Modified muse_mark to do explicit tail call elimination 
to support debug builds. With this modification the crash is 
eliminated. This "bugfix" affects only debug builds since the
compiler performs this optimization in release builds anyway.

-- extract from comments --
Compiler optimizations are turned off in debug builds and 
so it is easy to run into stack overflow conditions when
debugging with real data. Release build code should do
exactly what's done below to eliminate tail calls, but
I'm doing it by hand so that it takes effect in debug
builds as well. 
			
TODO: I'm not sure whether the explicit tail call 
elimination done here will cause the compiler to
generate less optimal code in the release build. That
remains to be measured.