0017177: Can't seek by clicking on seek bar
Added a flag called IsSeeking to MediaTimelineSlider and MagicMomentSlider.
Previously, click to seek is updated based on a timer checking for mouse capture event on the slider's thumb. The slider thumb position is updated immediately, and the timer will seek to the corresponding position if the thumb is still being mouse-captured.

However, because the timer is set a specific interval, sometime brief clicking might (mouse up) will cause the thumb to be "not mouse-captured" anymore by the time the timer reaches it. Hence the need to add another flag called IsSeeking to the slider, so that once slider's thumb is mouse captured, it's set totrue, and the client has the responsibility to set it back to false after checking.

Also updated slider for video dialog preview, music preview and voice over preview.