0018257: Access violation in thumbnail generation code
BUGFIX: There is a case where a Photo's Enter() may be called and
before it begins loading, its Leave() is called. This can happen if the
load is started asynchronously in the IO thread. When this happens,
some assumptions about the state of the IO thread break down.

Fixed this by making RenderEnvBase (which creates the IO thread) 
manages its initialization for the purpose of photo loading. 
This should get rid of any dependency between Leave() and 
LoadPhoto() inside ShowTime_Photo. 

TODO: This crash is hard to reproduce on Vista, but is conceivable.
In this commit, I've only verified that the common run cases have
not been altered by it. Need to verify on an XP machine tomorrow
where the crash seems to be more easily reproduced.

(Update - 18Mar08 - Gerry verified that this fix works on XP machines
and the crash is no longer reproducible).