#0018107 #0018398 EffectTrack is an array of effects within a muvee segment. It was created so that instead of each effect having to call the input media, the effects were bundled in an effectTrack array and thus reduced calls to the input media. EffectTrack with contains effects A and B calling input0 is better than effect A calling input0 and effect B calling input0. ( expand this scenario on a grand scale to appreciate it's effectiveness ).
However there was a bug in effectTrack inside the SetInput() function whereby each effect still ended calling the input media class. This has been removed and only the effectTrack now sets the input.

Also, inside the eval function of EffectTrack, when we need the effect at time t, we set the input. After we've evaled the effect, we set the input to NULL. That is because the Leave() function of the effect would cascade all the way down to the Leave() function of the input media. But this is already being done by the effectTrack, so we can safely remove the link.

This is a bug that affects all styles. It was particularly visible in frenzy because this style has lots of cut hints, flash hints and layers. These effects were inside different effectTracks. So at the end of the muvee, all the effects were calling the Leave() functions of the effectTracks beneath them which were in turn calling the Leave() functions beneath them etc etc. The Leave() function of showtime_photo ended up getting more than 500,000 calls ( I stopped counting there ) in a production of 4 pictures with music using frenzy style.