0015943: Duration of constructed muvee is different from the manual duration length set in "Personalize" dialog
RESOLVED: When manual duration is specified, the muvee duration is now computed based on
the title and credits duration, including transitions to body. So if you set duration to 20 seconds,
the total muvee duration will be 20s. For small durations, howeverm the muvee duration can end
up being longer. The duration you set is clamped at the lower end to include 5 seconds of body
and the title and credits.