0018824: Unable to use sound within an effect or transition
NOBUG: When specifying sound effects, you either need to specify the 'stop:' field
or the 'mstop:' field. Asserts will fail if both are left unspecified. This is
because not specifying 'stop:' means "play as much sound as there is, even if 
the sound goes beyond the effect boundary times." So 'mstop:' is therefore needed
to tell for how long the sound must play. OTOH, if 'mstop:' is omitted, it means
"play until you reach the given stop time". Therefore you need 'stop:' to be
specified.

A design alternative is to default 'stop:' to 1.0 - meaning stop the sound at
the end of the effect. This is not a very useful default since sound effects either
much shorter than a segment level effect's duration or bleed beyond the effect 
boundary. Therefore the 1.0 default is useful only when you're checking whether
sound effects work or not :) Since subtle sound  errors might get overlooked
if such a default is used, I prefer the explicit constraint of "either specify
stop: or mstop:".

In this commit, I've clarified a bunch of assertions to give clearer messages
about what was wrong.