0016556: Memory leak after save a muvee to file :: note 0047785
m_thread in <CMVCore::ProcMESTL2File::ProcMESTL2File()> is never released and causing leaks when saving muvee to a file. It's used to MVLOG an ending message when saving completes.

To fix the leaks simply, the use of ProcMESTL2File is commented out. The MVLOG statement is moved to SaveRenderer::Receive().