0018099: Anchorvale constructor refactoring
BUGFIX+REFACTOR: The algorithm to account for HL video and pictures together was getting
complicated, causing construction bugs for picture-only cases as well. So I've factored out
the code for the pictures only case and made it much simpler. The simplicity is possible because 
accounting for HL video prevents you from making some assumptions that can be made
when there is no video.