0018099: Anchorvale constructor refactoring
BUGFIX+REFACTOR: Rewrote MakeWith_AllPics_AllHLVideo to use an algorithm similar to
MakeWith_AllPics. The older algorithm was also trying to accommodate the fact that after
stretching/shrinking, the highlighted video might need to be played at a different speed
than before the act. This substantially complicated the algorithm. So I've now made a simplifying
assumption - that highlights will be played back at the speed at which they'll play when 
you use the "use all pictures and video" mode for the muvee - i.e the natural duration for
all pictures and video. This incongruence can usually be glossed over since it only applies when
making muvees with *only* pictures and highlighted video (i.e. no optional video). As it is,
the case is rare, and in most cases where the AllPics_AllHLVideo construction is performed, 
the result is discarded in favour of some other action such as drop pictures or use some optional
video.