0018099: Anchorvale constructor refactoring
FEATURE: Added support for dropping highlights partially. Prior to this 
change, a video highlight was either present in its entirety or absent
from the production. The new implementation will try to see if a muvee
can be made successfully with at least 1/2 of a highlight before dropping
it completely. 

Note that we may end up with music clipped because after dropping
a highlight we may not have enough material to fit to music. Therefore
the "music clippped" flag is suppresed explicitly when either pictures
are dropped or highlighted video is dropped.