0019088: Crash on "Make muvee" using sample media
BUGFIX: This is downright weird! The fix has nothing to do with music
or anything at all for that matter. In a loop, I was creating a local
variable of SegSpec structure, filling it up in the body of the loop
and at the end adding it to a vector. This bombed with the reported
error. I changed this scheme to first add a new entry to the vector
and fill the last entry in the body of the loop. This works fine! Is this
a compiler bug?