0019088: Crash on "Make muvee" using sample media
BUGFIX: Found the real problem and the solution.
The MakeWith_AllPics and MakeWith_AllPics_AllHLVideo modules
were both declaring and using a local structure named "SegSpec".
The problem is that they had different definitions but the same name
due to the illusion of locality. Once I renamed the one in MakeWith_AllPics_AllHLVideo
to "SegSpecPHLV" the crash vanishes. I suppose the linker should
have given at least a warning in this case, if not an error, but anyway.