0018739: [Constructor Tips] Style and Constructor Behaviour with Sample Media Files - Fit to Music
BUGFIX: Pictures were getting dropped in some simple cases that include video.
This bug was introduced when I clamped picture shrink factors to 1.0. Shrinking
is also used as an "error correction" mechanism in the old constructor and therefore
too much duration was some times being taken up by pictures. In this change,
I revert to the old explicit limits on stretching.

With this fix, if you make a muvee with sample media with the video file
appearing first, no picture wil be dropped.