0018165: Thumbnail generation – enhancement
Now thumbnail is generated in <FullScreenRenderer::Init()>. It is always the first frame of the muvee that's rendered to be the thumbnail.

Solution:
A preview window sized bitmap is first created. It is then resized to the requested size and saved to the requested path.

Advantages:
Doing so avoids running a third thumbnail renderer when making muvee, resulting less time for making muvee.

Todos:
1.) Logical change: the .mtn thumbnail file will be generated after preview starts. The old checking in <BatchListHelper::GenerateTimelineThumbnail()> will always fail, as the .mtn file won't be ready when <ShowTime_GenThumbnail()> returns.
	We should do the checking when the .mtn file is needed (e.g. when <Share Muvee> gets clicked).
2.) <BatchListHelper::GenerateTimelineThumbnail()> now only servers as a parameter passer. It simply passes the requested thumbnail size & path into MES and returns. Maybe we could just pass those info via the constructor of preview renderer?