0018362: Crash doing batch save with all 8 styles
Previously for saving, drv_open() & drv_close() are called in two threads and therefore the calls to CoInitialize() & CoInitialize() are not matched.

Deleting the CoInitialize() & CoInitialize() here simply works for both preview and saving, but we lack an explanation for the preview case, because COM won't be initialized then.
(For the saving case, COM is initialized in <CAMThread::InitialThreadProc()>, from <C:\DXSDK\Samples\C++\DirectShow\BaseClasses.VS2005\wxutil.cpp>.)

Now the COM library is initialized for FThreads, so we don't worry for the preview case.