#0016929: [TC12-01-02] Styles preview doesnt show up on machine without Windows Media Player 10 installed There would be no preview on systems that do not have WMP 10 and above. It would remain as a icon.
However, if the user were to make a muvee and proceed to select
another style, the style icon of the currently selected style
would disappear upon a selection of another style. See attached.
SOLUTION: I forgot to put in <exitaction> for data trigger when the style is used and no WMP 10 is installed. Put this in and the problem is solved.