0017555: Post construction 'tips' are not displayed or not displayed with correct values
BUGFIX: 

Problem - When pictures need to be stretched to fit to music, the suggested number
of pictures is way off. For example, with 22 pictures and 4 mins of music with Reflections
style, if you tweak the "speed" parameter to a little above the middle, you get 
suggested number of pictures to add as 70, 200, 300, etc. which are completely
unjustified. The maximum number of suggested pictures to add in this case is
the number of pictures put in - i.e. 22. (+/- depending on music characteristics).

Bug - the errRate * flex expression used in MakeWith_AllPics and MakeWith_AllPics_AllHLVideo
was using the *adjusted* value of flex instead of the original unadjusted value. "flex"
is the flexible duration of the pictures in the muvee, as computed without stretching or
shrinking.

Fix - Made flex a constant value so that it cannot be modified like it was in the previous
code and rewrote the duration expressions in terms of this const value. The modifying
expression 
  flex -= errRate * flex;
is now eliminated.

Comment - Very silly coding error. I must have been on anaesthesia when I wrote that!
Thanks to Karen for "pumping it up" and making the predicted number ridiculous enough
to be noticed.