0017555: Post construction 'tips' are not displayed or not displayed with correct values
BUGFIX: Fixed a formula error in determining the maximum amount of shrink
that can be applied in the pics-only case. Incidentally, this change will not
be visible *at all* in any style under any configuration!! It'll only
be visible when a style deliberately changes the "minimum factor to
which picture durations can be shrunk" to something other than 0.5.