0019358: [TCA04-03-02] Application would Crash During muvee Previewing
BUGFIX: Crash occurred at MESAudioGrabber::FillLastSegment().
The code was trying to allocate memory with -ve size.
Since -ve values wrap to large positive values when size
is treated as unsigned, this caused the heap to blow.
Fixed this by conditionally executing the allocation when
the size if +ve.

TODO: Sohrab to check the need for and correctness of the FillLastSegment()
function. I can't seem to understand what it is supposed to do,
except guess that it is somehow handling some left over data at the
end of audio playback.