0019430: Parameters.cpp(200) ASSERT FAILED: (t > m_segments.back().endt)
RESOLVED: The t == m_segments.back().endt condition is acceptable
only when t == 0.0. So I added a special case for this, making the given
animation value replace the "default" value that the animation was initialized
with.

TWEAKS: Made IntAnim and FloatAnim behave the same way, and made the
t == 0.0 condition robust rgd floating point errors by changing it to
t < 1e-5.