#0019442: Hang while seeking
BUG: Re-created the hang case and observed all the threads.
It showed that at the point of hang (after several several seeks),
there are two audio fragments being retained in the AudioFile
object when it is waiting for the real fragment of interest.
The fragment queue to the DirectShowVideo's audio code path is
only 2 fragments and therefore this caused a deadlock.

FIX: Before waiting for the required fragment, the AudioFile
object now returns all *other* buffers to the stream object.
This way, the stream will continue to decode the requested
fragment without blocking.