#0019287: Cancel saving issues: may hang app or causing runtime error on subsequent save
BUGFIX: The crashes were happening in ShowTime_VideoFile.cpp where a mutex's
reference count was ending up invalid. We tracked it down to the fact that
an FMutex object does Retain()/Release() for locked periods. After some
paper analysis, we figured that since FThread objects' task queues keep 
references to objects placed in the queue, there is no possibility for 
FMutexes to be destroyed while they are locked. Therefore we removed
the Retain()/Release() pairs in FMutex::Locker and FMutex::Unlocker.
The fact that FObject::Retain() and FObject::Release() now use
InterlockedIncrement() and InterlockedDecrement() should make this
behaviour correct for multi-core systems.