Resolved: #0017024: [TC27-01-01] Recorded voiceover segment does not contain any sound
This commit adds editable voice over track support so that when the user is
adding or removing voice over segments, the muvee doesn't need to be remade
in its entirety. Only the voice over part of the muvee can be remade. This
ensures quick response.

The constructor now maintains a reference to the last made muvee. Muvees are
always created with a voice over track, whether or not one was specified
by the user. This is in anticipation of the user's voice over addition.
When the user actually adds voice over segments, this dummy voice over
track is edited and filled with appropriate segment information. The edited
voice over track immediately becomes active in the renderer, thus avoiding
any need to remake the muvee.

By the time the user adds voice over segments, a renderer has already bee
created for the muvee made by the user. Therefore the voice over data has to
be communicated to the renderer by editing the voice over track directly.

This commit introduces a notion of "editable audio tracks" to mES. The
AudioTrackOp object now accepts two "parameters" - 'edit' for setting
its editability and 'clrs' for clearing all segments of an editable 
AudioTrackOp.

In order to update the voice over track, the constructor fetches
the last made voice over track (which can be a dummy) and first clears
it. It then adds all the voice over segments to the cleared track.
Since the muvee already holds a reference to the voice over track,
the new data is automatically reflected in the sound played back.

Some extra modifications to AudioTrackOp were needed in order to make this
happen - in particular, the interval of an AudioTrackOp is no longer fixed,
since it can contain editable tracks. Therefore an AudioTrackOp recomputes
its interval every time its Enter() function is called.

The VoiceOver dialog isn't fully functional at the moment. The volume
control sliders for music, sync sound and voice over are supposed to respond
in real time as the muvee plays back in the VO dialog. The controls only
take effect when the dialog is closed. Until the continuous control
feature is implemented, MESPreview will force set the audio volume
parameters every time StartPreview() is called - i.e. every time the
user hits the play button in the VO dialog.

In TLEngine.cpp, an assertion for audio part of the timeline being
non-zero in duration has been removed. The editable voice over 
track can have zero duration if the user hasn't specified any
voice over segments.