Resolved: #0019642: Picture dropped but video shown
The issue noted shows that there are cases where the old
constructor can fail to include all pictures, but still
include some video. Prior to this commit, the design is
such that the old constructor is invoked after the edge
case code has had a chance to examine the input material
to see whether all pictures and highlighted video can be
included in the output. Only if all pics and hl video 
can be included is the old constructor called upon to 
include some optional video as well.

This commit provides a fallback for the old constructor 
based on the edge case implementations. If, due to various 
cumulative anomalies, the old constructor failed to include 
all pictures and highlighted video, the appropriate
edge case implementation is used in its stead so that 
pictures are given the expected priority over video.

Misc: The change to muse_opcodes.h is to refine an assert
condition that raised a false alarm with the Reflections style
while testing this fix.