Resolved: #0019838: "Not Enough Video" when there *is* enough
PROBLEM: In this case, the StyleSpec determines that there is enough time
to show all pictures and some optional video and kicks the aP6
constructor into action. The aP6 constructor makes the muvee and finds
that it is just out of time to include the last picture and so drops it.
The safety net placed after the aP6 constructor determines that pictures
were dropped by video was shown and chooses to make a muvee with 
all the pictures. In order to fit to the music, the pictures need to be
time-stretched by a lot (> 2x) and therefore the stretch factor snaps back
to 1.0, creating a muvee that's shorter than the music duration. Hence
the "not enough video" message.

SOLUTION: Modified the aP6 constructor to *never* drop the pictures 
near the end of the input list. When it hits this critical condition, it extends
the muvee duration by as much as necessary to include all the pictures.
Though the muvee duration is extended, the music duration still remains
at the specified value, therefore the music will not repeat. So, in this case,
you get one picture at the end of the muvee with the music faded out.
Since the old constructor now never drops the last picture(s) in this
critical case, the safety net never kicks in.

NOTE: Maybe I should remove the safety net? I guess there is no harm
in leaving it in for some unforeseen circumstances.

NOTE: With this modification, you get "Video summarized, X clips dropped"
message.