0018739: [Constructor Tips] Style and Constructor Behaviour with Sample Media Files - Fit to Music
PROBLEM: As per note number 49264 - strips style with 5 pictures and 4 mins of video (sample video specified twice)
gives "not enough video" message - because the muvee duration is shorter than the music duration. 

DIAGNOSIS: The (aP6) constructor ran out of material by the time it reached the end of the muvee,
leaving the last segment with no material. 

SOLUTION: The constructor now keeps about 5 seconds of input "contributions" in reserve for 
use near the end of the muvee. If then the last segment is very short, it uses this extra contribution
to fill up the gap, extending the muvee duration if necessary by a few seconds (currently set to 2.5). 
When such an extension occurs near the end of music, the music is not repeated. With these
mods, the muvee truncation problem does not occur in cases such as the one found.

TODO: Check whether other cases where "not enough video" was reported have disappeared.