0019838: "Not Enough Video" when there *is* enough
PROBLEM: Same as described in note 51174.

DIAGNOSIS: Found a hole in the "repeat until all pictures are used" logic. The "somePicturesNotUsed"
flag was not computed in some cases near the end of the muvee, thereby terminating the muvee
before all pictures were used.

SOLUTION: Modified the loop so that somePicturesNotUsed is always computed (except when
running in KIA mode). Also the loop will continue to run if somePicturesNotUsed is true, irrespective
of other flags. I think this explicitly plugs all the holes. There are a few "break;"s within the 
construction loop which trigger when you reach the end of the media, but the loop body is
written such that it will not step over pictures - since all pictures are treated as "highlighted"
in aP6 terms. So it is not possible to run out of media and still not use all pictures. So these
are flagged as assertions in those "break;" blocks.

TODO: Check if this addresses Pete's project file.