#0019935 I am now commiting the change that removes those ugly shadows. It's the minimum possible change to create a presentable production. What I have done is display the captions in front of the effects. I have also set the captions alpha to 0.99 so that it blends nicely with the input. There is one down-side to this change. The captions will not move with the input. Essentially the caption is independent of the effect.
Why are we seeing the ugly dark shadows around our text?

That's because in "Pump it up", we have a lot of input layer replication and that causes the text to be also replicated. The shadow is usually quite low. But if you composite many shadows on top of one another, you get what we are seeing right now.